Change logic of army creation and change attack from double to int.

+++ SoftGame/Configuration/UnitsCoufiguration.cs
@@ -3,9 +3,9 @@
     public static class UnitsCoufiguration
     {
         public static int WarriorHealth = 50;
-        public static double WarriorAttack = 5;
+        public static int WarriorAttack = 5;
         
         public static int KnightHealth = 50;
-        public static double KnightAttack = 7;
+        public static int KnightAttack = 7;
     }
 }
